pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "GuilhermeSchreiber7", email = "guilhermeschreiber2007@gmail.com"},
]
dependencies = ["django>=5.1.6", "djangorestframework>=3.15.2"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
